pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "validators>=0.34.0",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.0.1",
    "ruff>=0.9.2",
]
